backend/lambda-functions/ai-handler/requirements.txt
reportlab==4.2.2
Pillow==10.4.0
fpdf2==2.7.6
fonttools==4.59.1
python-docx==1.1.2
PyPDF2==3.0.1
PyMuPDF==1.23.26
pdfplumber==0.10.3
boto3